Extracted gamespot dota reviews

+++ src/ricardo_crawlos/core/IReview.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public interface IReview
 {
     public Date getDataCreated();
-    public int getScore();
+    public double getScore();
     public String getComments();
     public int getGameID();
     public String getAuthor();
